# feat: introduce configuration system for serialization